Add dashboard support page and link powered-by footer
