+++ app/[locale]/dashboard/[guildId]/support/page.tsx
@@ -1,0 +1,9 @@
+import { SupportContent } from "@/components/support/support-content";
+
+export default function DashboardSupportPage() {
+  return (
+    <div className="min-h-screen bg-background">
+      <SupportContent compact />
+    </div>
+  );
+}
